backend/requirements.txt
fastapi
uvicorn
python-multipart
anthropic
SQLAlchemy
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
python-dotenv
PyPDF2
pdf2image
pytesseract
aiofiles
